requirements.txt
onf-parser
ai2-tango
beautifulsoup4
lxml
